requirements.txt
streamlit
pandas
numpy
joblib
folium
streamlit-folium
geopy
tensorflow
keras
pillow
